Add files via upload
This project was basically designed to help the visually impaired people to identify text and read. The camera module initially captures the image and gives it as an input to the model which then gives an output of the text written in the image by converting the text to audio.